% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbenchmark.R
\name{microbenchmark}
\alias{microbenchmark}
\title{A microbenchmarking function}
\usage{
microbenchmark(..., rep = 50, visual = FALSE)
}
\arguments{
\item{...}{Quoted expressions to analyse.}

\item{rep}{A number of time your code will be repeated, for better accuracy.}

\item{visual}{FALSE by default. If TRUE, returns a boxplot instead of values}
}
\value{
Either a list of ...length vectors of duration, or a boxplot of those vectors.
}
\description{
Use this to analyze your code and find what is the most efficient piece of code.
Duration is computed using system.time() and is composed of solely system time.
}
\examples{
microbenchmark("for (i in 1:1000000) {result <- sqrt(i)}",
"for (i in 1:1000000) {result <- i^2}", rep=10)

}
