% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moyenne.R
\name{moyenne}
\alias{moyenne}
\title{Moyenne d'un vecteur}
\usage{
moyenne(x)
}
\arguments{
\item{x}{A numericla vector}
}
\value{
returns the mean of x, minus NA
}
\description{
Une fonction pour faire une moyenne en enlevant les valeurs manquantes.
}
\examples{
moyenne(c(4,5))
}
